+++ apps/examples-react/bundle-size/xr/main.js
@@ -1,0 +1,3 @@
+import { createXrSessionController } from '@royal/react/xr';
+
+globalThis.__royalCreateXrSessionController = createXrSessionController;